requirements.txt
networkx>=2.3
biopython>=1.78